third_party/WebKit/LayoutTests/typedcssom/cssRotation-expected.txt

This is a testharness.js-based test.
PASS Invalid arguments to constructor should throw 
PASS x, y, z and angle values are correct for rotate(0deg) 
PASS x, y, z and angle values are correct for rotate(10deg) 
PASS x, y, z and angle values are correct for rotate(-21deg) 
PASS x, y, z and angle values are correct for rotate(3.2deg) 
PASS x, y, z and angle values are correct for rotate3d(0, 0, 1, 90deg) 
PASS x, y, z and angle values are correct for rotate3d(2.7, -3, 4.4, 0deg) 
PASS x, y, z and angle values are correct for rotate3d(2, 3, 4, 10deg) 
PASS x, y, z and angle values are correct for rotate3d(2, 3.7, -4, -1.2deg) 
PASS x, y, z and angle values are correct for rotate3d(1, 0, 0, 0.5turn) 
FAIL is2D value is correct for rotate(0deg) assert_equals: expected true but got false
FAIL is2D value is correct for rotate(10deg) assert_equals: expected true but got false
FAIL is2D value is correct for rotate(-21deg) assert_equals: expected true but got false
FAIL is2D value is correct for rotate(3.2deg) assert_equals: expected true but got false
PASS is2D value is correct for rotate3d(0, 0, 1, 90deg) 
PASS is2D value is correct for rotate3d(2.7, -3, 4.4, 0deg) 
PASS is2D value is correct for rotate3d(2, 3, 4, 10deg) 
PASS is2D value is correct for rotate3d(2, 3.7, -4, -1.2deg) 
PASS is2D value is correct for rotate3d(1, 0, 0, 0.5turn) 
FAIL toString value is correct for rotate(0deg) assert_equals: expected "rotate(0deg)" but got ""
FAIL toString value is correct for rotate(10deg) assert_equals: expected "rotate(10deg)" but got ""
FAIL toString value is correct for rotate(-21deg) assert_equals: expected "rotate(-21deg)" but got ""
FAIL toString value is correct for rotate(3.2deg) assert_equals: expected "rotate(3.2deg)" but got ""
FAIL toString value is correct for rotate3d(0, 0, 1, 90deg) assert_equals: expected "rotate3d(0, 0, 1, 90deg)" but got ""
FAIL toString value is correct for rotate3d(2.7, -3, 4.4, 0deg) assert_equals: expected "rotate3d(2.7, -3, 4.4, 0deg)" but got ""
FAIL toString value is correct for rotate3d(2, 3, 4, 10deg) assert_equals: expected "rotate3d(2, 3, 4, 10deg)" but got ""
FAIL toString value is correct for rotate3d(2, 3.7, -4, -1.2deg) assert_equals: expected "rotate3d(2, 3.7, -4, -1.2deg)" but got ""
FAIL toString value is correct for rotate3d(1, 0, 0, 0.5turn) assert_equals: expected "rotate3d(1, 0, 0, 0.5turn)" but got ""
FAIL toMatrix works for rotate(0deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate(10deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate(-21deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate(3.2deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate3d(0, 0, 1, 90deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate3d(2.7, -3, 4.4, 0deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate3d(2, 3, 4, 10deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate3d(2, 3.7, -4, -1.2deg) assert_equals: is2D expected false but got true
FAIL toMatrix works for rotate3d(1, 0, 0, 0.5turn) assert_equals: is2D expected false but got true
Harness: the test ran to completion.
